add optional ssl support
